epistemic-risk-detector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epistemic-risk-detector"
version = "0.1.0"
description = "Research-grade prototype for inspecting epistemic risk in LLM outputs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Epistemic Risk Detector Contributors"}
]
keywords = ["llm", "hallucination", "epistemic", "risk", "nlp", "ai", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "numpy>=1.24",
    "sentence-transformers>=2.2",
    "sqlite-vec>=0.1.0",
    "openai>=1.0",
    "anthropic>=0.18",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]
local = [
    "llama-cpp-python>=0.2",
]

[project.scripts]
epistemic-risk = "src.cli:main"

[project.urls]
Repository = "https://github.com/your-org/epistemic-risk-detector"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
